refactor: create index.ts files for component exports
