+++ src/components/account/index.ts
@@ -1,0 +1,1 @@
+export { default as AccountOverview } from './AccountOverview';
