+++ src/pages/MyAccount.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import { Header, Footer } from '@components/common';
-import AccountOverview from '@components/account/AccountOverview';
+import { AccountOverview } from '@components/account';
 
 export default function MyAccount() {
   return (
     <>
       <Header />
-      <AccountOverview />
+      <main>
+        <AccountOverview />
+      </main>
       <Footer />
     </>
   );
