+++ src/pages/Theme.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import SectionHeader from '@components/common/SectionHeader';
-import GoodsItemList from '@components/theme/GoodsItemList';
-import { Header, Footer } from '@/components/common';
+import { Header, Footer, SectionHeader } from '@components/common';
+import { GoodsItemList } from '@components/theme';
 
 export default function Theme() {
   return (
